% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_aic.R
\name{approx_aic}
\alias{approx_aic}
\title{Approximate AIC Calculation}
\usage{
approx_aic(model, nsim = 1000)
}
\arguments{
\item{model}{object; generated from \code{\link[mergm]{mergm}}, or
generated from \code{\link[ergm]{ergm}}.}

\item{nsim}{count; number of network simulations needed for the calculation
of the approx. AIC value. Default = 1000}
}
\description{
Function to calculate an approximate AIC value for both mERGM and ERGM.
\code{\link[mergm]} and  ERGM \code{\link[ergm]} for model comparison.
}
\details{
This function returns an approx. AIC value which can be used for
model comparison.
}
\examples{
\dontrun{
# Load the karate club social network of Zachary (1977)
data("zach", package = "ergm.count")
# Seed for reproducity
set.seed(2410)
# Fit a model to measur the propensity to form edges, 2-stars and
# capture the unobserved nodal heterogeneity, allowing 10 iteration steps.
mod <- mergm(zach ~ edges + kstar(2), 10)
# Get the AIC value
approx_aic(mod)
}

}
\seealso{
\code{\link[mergm]{mergm}}.

\code{\link[mergm]{convergence}}.
}
