% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergm.R
\name{mergm}
\alias{mergm}
\title{Iterative Estimation of Mixed Exponential Random Graph Models with Nodal
Random Effects.}
\usage{
mergm(formula, iter)
}
\arguments{
\item{formula}{formula; an \code{R} formula object, of the form
<network> ~ <model terms> where <network> is a \code{\link[network]{network}}
object and <model terms> are \code{\link[ergm]{ergm.terms}}. The term
\code{sociality(1:nodes)} is added automatically to the formula argument.}

\item{iter}{count; number of iterations used for iterative estimation
of the model coefficients and the random effects, 10 iterations are usually
enough.}
}
\description{
Function to fit ERGMs with nodal random effects. Estimation is carried out
combining the stepping algorithm (see \code{\link[ergm]{control.ergm}}) for
the fixed parameters and pseudo-likelihood (see \code{\link[mgcv]{bam}})
for the nodal random effects.
}
\details{
This implemented algorithm can only handle undirected networks.
}
\examples{
\dontrun{
# Load the karate club social network of Zachary (1977)
data("zach", package = "ergm.count")
# Seed for reproducity
set.seed(2410)
# Fit a model to measur the propensity to form edges, 2-stars and
# capture the unobserved nodal heterogeneity, allowing 10 iteration steps.
mod <- mergm(zach ~ edges + kstar(2), 10)
# GOF of the model
gof(mod$model)
plot(gof(mod$model))
# Summary of the model
summary(mod$model)
}
}
\references{
Kevork, S. and Kaeurmann, G. (2022).
Iterative Estimation of Mixed Exponential Random Graph Models with Nodal
Random Effects. Network Science 9, 478-498. https://doi.org/10.1017/nws.2021.22.
}
\author{
Sevag Kevork
}
